requirements.txt
agno
phi
groq
httpx
beautifulsoup4
python-dotenv
supabase
apscheduler
python-telegram-bot
pydantic
sqlalchemy
pytest
requests